% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shirtShuffler.R
\name{shuffleShirts}
\alias{shuffleShirts}
\title{Choose my shirts}
\usage{
shuffleShirts(rawShirts = system.file("rawShirts", package = "shirtshufflr"),
  output = file.path(tempdir(), "shuffledShirts"))
}
\arguments{
\item{rawShirts}{file path to raw shirt files}

\item{output}{file path for output files}
}
\value{
character vector of shirt order filenames
}
\description{
Choose which shirt I will wear each weekday
}
\details{
Raw shirt patterns are found in the directory given by
\code{rawShirts}. This function randomly selects 5 of the patterns,
and copies files names for each weekday into the directory given by
\code{output}. Just open the image for each day of the week to
find your shirt surprise.

You can have as many shirts as you like in \file{rawShirts}, but
you should probably have at least five or the sample will fail.

This function also assumes your shirt images are JPG files, so
probably just make sure they are JPG files.
}
